Solve parameter extraction
