+++ backend/app/core/authz/authorization.py
@@ -4,23 +4,14 @@ from typing import cast
 
 import casbin_async_sqlalchemy_adapter as sqlalchemy_adapter
 from casbin import AsyncEnforcer
-from fastapi import Depends, HTTPException, status
+from fastapi import Depends, HTTPException, Request, status
 
 from app.core.auth.auth import get_authenticated_user
 from app.database import database
 from app.users.schema import User
+from app.utils.singleton import Singleton
 
 from .actions import AuthorizationAction
-from .parameters import AuthorizationDep
-
-
-class Singleton(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
 
 
 class Authorization(metaclass=Singleton):
@@ -42,20 +33,35 @@ class Authorization(metaclass=Singleton):
 
     @classmethod
     def enforce(
-        cls, action: AuthorizationAction
-    ) -> Callable[[AuthorizationDep, User], None]:
+        cls,
+        action: AuthorizationAction,
+        object_id: str = "object_id",
+        domain: str = "domain",
+    ) -> Callable[[Request, User], None]:
         def inner(
-            params: AuthorizationDep,
+            request: Request,
             user: User = Depends(get_authenticated_user),
         ) -> None:
+            obj = cls.resolve_parameter(request, object_id)
+            dom = cls.resolve_parameter(request, domain)
+
             return cls()._enforce(
                 sub=str(user.id),
-                dom=params.domain,
-                obj=params.object_id,
+                dom=dom,
+                obj=obj,
                 act=action,
             )
 
         return inner
+
+    @staticmethod
+    def resolve_parameter(request: Request, key: str, default: str = "*") -> str:
+        if (result := request.query_params.get(key)) is not None:
+            return cast(str, result)
+        if (result := request.path_params.get(key)) is not None:
+            return cast(str, result)
+
+        return default
 
     def _enforce(self, *, sub: str, dom: str, obj: str, act: str) -> None:
         enforcer: AsyncEnforcer = self.enforcer
